remove duolication of const DRUG_ROUTE_VALUE_TO_LABEL
