+++ packages/constants/src/medications.ts
@@ -1,0 +1,15 @@
+export const DRUG_ROUTE_VALUE_TO_LABEL = {
+  dermal: 'Dermal',
+  ear: 'Ear',
+  eye: 'Eye',
+  intramuscular: 'IM',
+  intravenous: 'IV',
+  inhaled: 'Inhaled',
+  nasal: 'Nasal',
+  oral: 'Oral',
+  rectal: 'Rectal',
+  subcutaneous: 'S/C',
+  sublingual: 'Sublingual',
+  topical: 'Topical',
+  vaginal: 'Vaginal',
+};
